Added basic windows support for electron.
There is no make file. You need to run 'cl.exe *.c /oelectron.exe'
and it will spit out the appropriate executable. I tested it on
windows 7 SP1 x64. I had to make some changes to the code to get it
to compile (and made fixes in the process.) Unfortunately, windows
is different enough that supporting shell variables is a pain in
the ass and as such isn't enabled at this point.