Game consoles are no longer supported.
It should be up to an accompanying C program to interact with electron and
handle the operating system specific features such as drawing and such.
It should not be up to the expert system at all in any specific fashion.

If support for these consoles are necessary then electron should be modified
specially for them not included in the standard distribution.

I also removed support for architecture version detection as well.